refactor: refactoring the giftList page components
